No Preference should be # of voters.  Fixes #7.
"No Preference" should be the number of *voters*
who expressed no preference between any of the candidates
in what I call a "preference round".  It was actually
printing the number of individual head-to-head contests
in the round that had no preference, and there could be
more than one on a single ballot.