2.1!  Change to verbosity for permuted tiebreaker.
Change to `verbosity`.  If `verbosity` is 1, the permuted
tiebreaker doesn't print any output at initialization time.
Instead, if there's a tiebreaker, it prints the pre-permuted
list of candidates immediately before printing the results of
the tiebreaker.  Note that this doesn't change when the
initialization is done; when the the list of candidates is
randomly permuted by **starvote**, that's still done *before*
running the election.  If `verbosity` is 2 or higher,
the permuted tiebreaker initialization prints its output at
initialization time as before.

There were some other bug fixes and one more feature request
in this release; see README.md (or git log) for the full list.