Tighten up print-blank-line-before-header logic.
Now it only prints a blank line before a header if we print
any *visible* characters.  Before any line would work, even
whitespace-only lines.

I'm not sure that can actually happen; I believe I always set
a heading on output, and the heading is always non-blank.
The new if statement with a "pragma no cover" is currently
never run in the test suite, and I'm not sure it ever *can*
be run given the current implementation.  But in case I ever
change my output format such that I don't print headings or
something, the output logic is ready, fingers crossed.